+++ SpicesEngine/src/Render/Vulkan/VulkanSwapChain.cpp
@@ -138,7 +138,7 @@ namespace Spices {
 		* @breif Create SwapChain.
 		*/
 		VK_CHECK(vkCreateSwapchainKHR(m_VulkanState.m_Device, &createInfo, nullptr, &m_VulkanState.m_SwapChain));
-		VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SWAPCHAIN_KHR, m_VulkanState.m_SwapChain, m_VulkanState.m_Device, "SpicesEngineSwapChainKHR");
+		VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SWAPCHAIN_KHR, (uint64_t)m_VulkanState.m_SwapChain, m_VulkanState.m_Device, "SpicesEngineSwapChainKHR");
 
 		/**
 		* @brief Get Swapchain images created by SwapChain automatically.
@@ -176,7 +176,7 @@ namespace Spices {
 				info.subresourceRange.baseArrayLayer = 0;
 				info.subresourceRange.layerCount     = 1;
 
-				VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_IMAGE, m_VulkanState.m_SwapChainImages[i], m_VulkanState.m_Device, "SwapChainImage");
+				VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_IMAGE, (uint64_t)m_VulkanState.m_SwapChainImages[i], m_VulkanState.m_Device, "SwapChainImage");
 				VK_CHECK(vkCreateImageView(m_VulkanState.m_Device, &info, nullptr, &m_VulkanState.m_SwapChainImageViews[i]))
 			}
 
@@ -265,24 +265,24 @@ namespace Spices {
 		{
 			// Graphic SyncObkects.
 			VK_CHECK(vkCreateSemaphore(m_VulkanState.m_Device, &semaphoreInfo, nullptr, &m_VulkanState.m_GraphicImageSemaphore[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, m_VulkanState.m_GraphicImageSemaphore[i], m_VulkanState.m_Device, "GraphicImageSemaphore");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, (uint64_t)m_VulkanState.m_GraphicImageSemaphore[i], m_VulkanState.m_Device, "GraphicImageSemaphore");
 
 			VK_CHECK(vkCreateSemaphore(m_VulkanState.m_Device, &semaphoreInfo, nullptr, &m_VulkanState.m_GraphicQueueSemaphore[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, m_VulkanState.m_GraphicQueueSemaphore[i], m_VulkanState.m_Device, "GraphicQueueSemaphore");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, (uint64_t)m_VulkanState.m_GraphicQueueSemaphore[i], m_VulkanState.m_Device, "GraphicQueueSemaphore");
 
 			VK_CHECK(vkCreateFence(m_VulkanState.m_Device, &fenceInfo, nullptr, &m_VulkanState.m_GraphicFence[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_FENCE, m_VulkanState.m_GraphicFence[i], m_VulkanState.m_Device, "GraphicFence");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_FENCE, (uint64_t)m_VulkanState.m_GraphicFence[i], m_VulkanState.m_Device, "GraphicFence");
 
 
 			// Compute SyncObkects.
 			VK_CHECK(vkCreateSemaphore(m_VulkanState.m_Device, &semaphoreInfo, nullptr, &m_VulkanState.m_ComputeImageSemaphore[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, m_VulkanState.m_ComputeImageSemaphore[i], m_VulkanState.m_Device, "ComputeImageSemaphore");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, (uint64_t)m_VulkanState.m_ComputeImageSemaphore[i], m_VulkanState.m_Device, "ComputeImageSemaphore");
 
 			VK_CHECK(vkCreateSemaphore(m_VulkanState.m_Device, &semaphoreInfo, nullptr, &m_VulkanState.m_ComputeQueueSemaphore[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, m_VulkanState.m_ComputeQueueSemaphore[i], m_VulkanState.m_Device, "ComputeQueueSemaphore");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_SEMAPHORE, (uint64_t)m_VulkanState.m_ComputeQueueSemaphore[i], m_VulkanState.m_Device, "ComputeQueueSemaphore");
 
 			VK_CHECK(vkCreateFence(m_VulkanState.m_Device, &fenceInfo, nullptr, &m_VulkanState.m_ComputeFence[i]))
-			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_FENCE, m_VulkanState.m_ComputeFence[i], m_VulkanState.m_Device, "ComputeFence");
+			VulkanDebugUtils::SetObjectName(VK_OBJECT_TYPE_FENCE, (uint64_t)m_VulkanState.m_ComputeFence[i], m_VulkanState.m_Device, "ComputeFence");
 		}
 	}
